silence warnings

+++ io/level-io.c
@@ -4,19 +4,16 @@ static void process_row(char *line, Jagged_Row *row);
 
 // Main function to read and create the grid
 // TODO ! Strip empty final rows if there are any
-extern Jagged_Grid *read_grid_csv_file(const char *filename)
-{
+extern Jagged_Grid *read_grid_csv_file(const char *filename) {
   FILE *file = fopen(filename, "r");
-  if (!file)
-  {
+  if (!file) {
     fprintf(stderr, "Could not open file %s\n", filename);
     return NULL;
   }
 
   // Create the grid structure
   Jagged_Grid *grid = malloc(sizeof(Jagged_Grid));
-  if (!grid)
-  {
+  if (!grid) {
     fclose(file);
     return NULL;
   }
@@ -24,8 +21,7 @@ extern Jagged_Grid *read_grid_csv_file(const char *filename)
   // First pass: count the number of rows
   grid->length = 0;
   char c;
-  while ((c = fgetc(file)) != EOF)
-  {
+  while ((c = fgetc(file)) != EOF) {
     if (c == '\n')
       grid->length++;
   }
@@ -34,8 +30,7 @@ extern Jagged_Grid *read_grid_csv_file(const char *filename)
 
   // Allocate rows array
   grid->rows = malloc(grid->length * sizeof(Jagged_Row));
-  if (!grid->rows)
-  {
+  if (!grid->rows) {
     free(grid);
     fclose(file);
     return NULL;
@@ -45,14 +40,14 @@ extern Jagged_Grid *read_grid_csv_file(const char *filename)
   rewind(file);
 
   // Buffer for reading lines
-  char *line = NULL;
-  size_t len = 0;
+  char   *line = NULL;
+  size_t  len  = 0;
   ssize_t read;
-  int row_index = 0;
+  int     row_index = 0;
 
   // Read and process each line
-  while ((read = getline(&line, &len, file)) != -1 && row_index < grid->length)
-  {
+  while ((read = getline(&line, &len, file)) != -1 &&
+         (size_t)row_index < grid->length) {
     process_row(line, &grid->rows[row_index]);
     row_index++;
   }
@@ -62,19 +57,17 @@ extern Jagged_Grid *read_grid_csv_file(const char *filename)
   return grid;
 }
 
-static void process_row(char *line, Jagged_Row *row)
-{
+static void process_row(char *line, Jagged_Row *row) {
   int last_content = -1;
-  int current_pos = 0;
-  int has_content = 0;
+  int current_pos  = 0;
+  int has_content  = 0;
 
   char *line_copy = strdup(line);
-  char *pos = line_copy;
+  char *pos       = line_copy;
   char *token;
 
   // First pass: find last non-empty position
-  while ((token = strsep(&pos, ",")) != NULL)
-  {
+  while ((token = strsep(&pos, ",")) != NULL) {
     // Trim whitespace and newline
     while (*token && (isspace(*token) || *token == '\n'))
       token++;
@@ -83,39 +76,36 @@ static void process_row(char *line, Jagged_Row *row)
       end--;
     *end = '\0';
 
-    if (*token)
-    {
-      has_content = 1;
+    if (*token) {
+      has_content  = 1;
       last_content = current_pos;
     }
     current_pos++;
   }
   free(line_copy);
 
-  if (!has_content)
-  {
-    row->length = 0;
+  if (!has_content) {
+    row->length             = 0;
     row->world_object_names = NULL;
     printf("Setting to NULL line 99");
     return;
   }
 
-  row->length = last_content + 1;
+  row->length             = last_content + 1;
   row->world_object_names = malloc(row->length * sizeof(Object_Name));
-  if (!row->world_object_names)
-  {
+  if (!row->world_object_names) {
     printf("Setting to length 0 line 107");
     row->length = 0;
     return;
   }
 
   // Second pass: store strings
-  line_copy = strdup(line);
-  pos = line_copy;
+  line_copy   = strdup(line);
+  pos         = line_copy;
   current_pos = 0;
 
-  while ((token = strsep(&pos, ",")) != NULL && current_pos < row->length)
-  {
+  while ((token = strsep(&pos, ",")) != NULL &&
+         (size_t)current_pos < row->length) {
     // Trim whitespace and newline
     while (*token && (isspace(*token) || *token == '\n'))
       token++;
@@ -124,26 +114,21 @@ static void process_row(char *line, Jagged_Row *row)
       end--;
     *end = '\0';
 
-    if (!*token)
-    {
+    if (!*token) {
       row->world_object_names[current_pos] = strdup("EMPTY");
-    }
-    else
-    {
+    } else {
       row->world_object_names[current_pos] = strdup(token);
     }
 
-    if (!row->world_object_names[current_pos])
-    {
+    if (!row->world_object_names[current_pos]) {
       // Memory allocation failed, clean up
-      for (size_t i = 0; i < current_pos; i++)
-      {
+      for (size_t i = 0; i < (size_t)current_pos; i++) {
         printf("Memory fail line 141");
         free(row->world_object_names[i]);
       }
       free(row->world_object_names);
       row->world_object_names = NULL;
-      row->length = 0;
+      row->length             = 0;
       free(line_copy);
       return;
     }
@@ -153,17 +138,13 @@ static void process_row(char *line, Jagged_Row *row)
   free(line_copy);
 }
 
-extern void free_jagged_grid(Jagged_Grid *grid)
-{
+extern void free_jagged_grid(Jagged_Grid *grid) {
   if (!grid)
     return;
 
-  for (size_t i = 0; i < grid->length; i++)
-  {
-    if (grid->rows[i].world_object_names)
-    {
-      for (size_t j = 0; j < grid->rows[i].length; j++)
-      {
+  for (size_t i = 0; i < grid->length; i++) {
+    if (grid->rows[i].world_object_names) {
+      for (size_t j = 0; j < grid->rows[i].length; j++) {
         free(grid->rows[i].world_object_names[j]);
       }
       free(grid->rows[i].world_object_names);
@@ -173,26 +154,20 @@ extern void free_jagged_grid(Jagged_Grid *grid)
   free(grid);
 }
 
-extern void print_jagged_grid(const Jagged_Grid *grid)
-{
-  if (!grid)
-  {
+extern void print_jagged_grid(const Jagged_Grid *grid) {
+  if (!grid) {
     printf("Grid is NULL\n");
     return;
   }
 
   printf("Grid has %zu rows:\n", grid->length);
-  for (size_t i = 0; i < grid->length; i++)
-  {
+  for (size_t i = 0; i < grid->length; i++) {
     printf("Row %zu: length=%zu, elements=", i, grid->rows[i].length);
 
-    if (grid->rows[i].world_object_names)
-    {
-      for (size_t j = 0; j < grid->rows[i].length; j++)
-      {
+    if (grid->rows[i].world_object_names) {
+      for (size_t j = 0; j < grid->rows[i].length; j++) {
         printf("%s", grid->rows[i].world_object_names[j]);
-        if (j < grid->rows[i].length - 1)
-        {
+        if (j < grid->rows[i].length - 1) {
           printf(",");
         }
       }
