+++ io/read-manifest.c
@@ -1,10 +1,8 @@
 #include "./read-manifest.h"
 
-extern char *read_asset_manifest_file(const char *filename)
-{
+extern char *read_asset_manifest_file(const char *filename) {
   FILE *file = fopen(filename, "r");
-  if (!file)
-  {
+  if (!file) {
     fprintf(stderr, "Could not open file %s\n", filename);
     return NULL;
   }
@@ -14,16 +12,14 @@ extern char *read_asset_manifest_file(const char *filename)
   fseek(file, 0, SEEK_SET);
 
   char *buffer = malloc(file_size + 1); // extra for null terminator
-  if (!buffer)
-  {
+  if (!buffer) {
     fprintf(stderr, "Memory allocation failed for manifest\n");
     fclose(file);
     return NULL;
   }
 
   size_t read_size = fread(buffer, 1, file_size, file);
-  if (read_size != file_size)
-  {
+  if (read_size != (size_t)file_size) {
     fprintf(stderr, "Failed to read file %s\n", filename);
     free(buffer);
     fclose(file);
